# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  handlers.CreateReviewRequest:
    properties:
      contains_spoilers:
        example: false
        type: boolean
      content:
        example: Great movie!
        type: string
      rating:
        example: 4.5
        type: number
    required:
    - rating
    type: object
...
  handlers.MovieSummaryResponse:
    properties:
      date:
        example: "1999-10-15"
        type: string
      duskforge_rating:
        example: 4.5
        type: number
      id:
        example: 550
        type: integer
      name:
        example: Fight Club
        type: string
      poster:
        example: /pB8BM7pdSp6B6Ih7QZ4DrQ3PmJK.jpg
        type: string
      tmdb_rating:
        example: 4.3
        type: number
    type: object
...
  handlers.ReviewResponse:
    properties:
      comment_count:
        example: 5
        type: integer
      contains_spoilers:
        example: false
        type: boolean
      content:
        example: Great movie!
        type: string
      created_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      like_count:
        example: 12
        type: integer
      liked_by_user:
        example: false
        type: boolean
      movie:
        $ref: '#/definitions/handlers.MovieSummaryResponse'
      rating:
        example: 4.5
        type: number
      updated_at:
        example: "2024-01-15T10:30:00Z"
        type: string
      user:
        $ref: '#/definitions/handlers.UserSummary'
    type: object
...
  handlers.UpdateReviewRequest:
    properties:
      contains_spoilers:
        example: true
        type: boolean
      content:
        example: Updated review
        type: string
      rating:
        example: 4
        type: number
    type: object
...
  handlers.UserSummary:
    properties:
      avatar_url:
        example: https://example.com/avatar.jpg
        type: string
      id:
        example: 660e8400-e29b-41d4-a716-446655440000
        type: string
      username:
        example: johndoe
        type: string
    type: object
...
  response.ErrorInfo:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
  response.PaginatedResponse:
    properties:
      data: {}
      pagination:
        $ref: '#/definitions/response.Pagination'
      success:
        type: boolean
    type: object
  response.Pagination:
    properties:
      limit:
        type: integer
      offset:
        type: integer
      total:
        type: integer
    type: object
  response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/response.ErrorInfo'
      success:
        type: boolean
    type: object
info:
  contact:
    email: support@duskforge.io
    name: API Support
    url: https://github.com/duskforge/api
  description: A movie collection and review API powered by TMDB
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: Duskforge API
  version: "1.0"
paths:
...
  /movies/{id}/reviews:
    get:
      description: List all reviews for a movie by TMDB ID with pagination and sorting.
        Only reviews with content are returned. If authenticated, the logged-in user's
        own review is excluded and reviews by blocked users are filtered out. Reviews
        by banned users are hidden for non-admin callers.
      parameters:
      - description: TMDB movie ID
        in: path
        name: id
        required: true
        type: integer
      - default: 0
        description: Offset for pagination
        in: query
        name: offset
        type: integer
      - default: 20
        description: Limit for pagination
        in: query
        name: limit
        type: integer
      - default: -likes
        description: Sort field with direction prefix (+asc, -desc)
        enum:
        - +likes
        - -likes
        - +rating
        - -rating
        - +created_at
        - -created_at
        in: query
        name: sort
        type: string
      - description: Language code (e.g., en, fr)
        in: header
        name: Accept-Language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: List of reviews
          schema:
            allOf:
            - $ref: '#/definitions/response.PaginatedResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/handlers.ReviewResponse'
                  type: array
              type: object
        "400":
          description: Invalid movie ID
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get reviews for a movie
      tags:
      - reviews
    post:
      consumes:
      - application/json
      description: Create a review for a movie by TMDB ID
      parameters:
      - description: TMDB movie ID
        in: path
        name: id
        required: true
        type: integer
      - description: Review details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateReviewRequest'
      - description: Language code (e.g., en, fr)
        in: header
        name: Accept-Language
        type: string
      produces:
      - application/json
      responses:
        "201":
          description: Review created
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.ReviewResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Review already exists for this movie
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create a review
      tags:
      - reviews
...
  /reviews/{reviewId}:
    delete:
      description: Delete your own review
      parameters:
      - description: Review ID
        format: uuid
        in: path
        name: reviewId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Review deleted
        "400":
          description: Invalid review ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Review not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Delete a review
      tags:
      - reviews
    get:
      description: Get a single review by its ID. Returns 404 if the review author
        is banned (non-admin callers). Returns 403 if there is a block between the
        authenticated user and the review author.
      parameters:
      - description: Review ID
        format: uuid
        in: path
        name: reviewId
        required: true
        type: string
      - description: Language code (e.g., en, fr)
        in: header
        name: Accept-Language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Review details
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.ReviewResponse'
              type: object
        "400":
          description: Invalid review ID
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: User blocked
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Review not found or author banned
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get a review by ID
      tags:
      - reviews
    patch:
      consumes:
      - application/json
      description: Update your own review
      parameters:
      - description: Review ID
        format: uuid
        in: path
        name: reviewId
        required: true
        type: string
      - description: Fields to update
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateReviewRequest'
      - description: Language code (e.g., en, fr)
        in: header
        name: Accept-Language
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Updated review
          schema:
            allOf:
            - $ref: '#/definitions/response.Response'
            - properties:
                data:
                  $ref: '#/definitions/handlers.ReviewResponse'
              type: object
        "400":
          description: Invalid request body
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Review not found
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update a review
      tags:
      - reviews
...
  /reviews/{reviewId}/like:
    delete:
      description: Remove a like from a review. Returns 403 if there is a block between
        the authenticated user and the review author.
      parameters:
      - description: Review ID
        format: uuid
        in: path
        name: reviewId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Review unliked
        "400":
          description: Invalid review ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: User blocked
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Review not found or not liked
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Unlike a review
      tags:
      - reviews
    post:
      description: Like a review. Returns 403 if there is a block between the authenticated
        user and the review author.
      parameters:
      - description: Review ID
        format: uuid
        in: path
        name: reviewId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Review liked
        "400":
          description: Invalid review ID
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: User blocked
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Review not found
          schema:
            $ref: '#/definitions/response.Response'
        "409":
          description: Already liked
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Like a review
      tags:
      - reviews
...
securityDefinitions:
  BearerAuth:
    description: 'Enter your bearer token in the format: Bearer {token}'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"